Add files via upload
To create a “Weather Forecast website” which provides the information regarding the weather at any place using an API. This website development includes the usage of an API, which fetches the real time data of the weather at any place and forecasts the weather for the upcoming 5 days. Automatic location detection is to be enabled in this project to improve the versatility of the website.